feat: add admin layout client and manual test
